requirements.txt
torch==2.3.0.dev20231214
Pillow==10.0.1
ttkbootstrap==1.10.1
diffusers==0.24.0
